Strange Cost for when mapw is 1.

+++ algorithm.h
@@ -4,11 +4,11 @@
 typedef struct path_cost
 {
 	int pos[2];
-	int cost;
-	int distance;
+	float cost;
+	float distance;
 	struct path_cost *next, *prev;
 } PathCost;
 
-PathCost *a_star_search(int init[2], int goal[2], int *distance);
+PathCost *a_star_search(int init[2], int goal[2], float *distance);
 
 #endif /* ALGORITHM_H */
